ci: Don't build again for creating references
Enable the docs in the main build, then inherit the artifacts. It should
result in less CPU resources being used.

GitLab doesn't support multiple artifacts per job, so we still have to
download the whole build, as the build job can't hand out just the
references, and it seems to be planned as a premium-only feature sadly:
https://gitlab.com/gitlab-org/gitlab/-/issues/18744

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/3816>