backends/native: Only warn about cursor sizes once
Because if the current theme has exceeded the dimensions of
`DRM_CAP_CURSOR_WIDTH/HEIGHT` then the warning is just going to repeat
every time the cursor changes. We still fall back to software cursors
just fine so it's not important to repeat the warning.

Also the theme cursor size is not "Invalid" so we change the message to
match that of 47 main.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/3597