cogl/pango: Remove private cogl headers usage
As we will be merging CoglPango inside Clutter, so prepare for that
Note this removes the Cogl shared atlas debug usage, as it is private to Cogl

The usage of cogl_c_args was replaced with cogl_debug_c_args to avoid COGL_COMPILATION
pre-moving to clutter. The whole meson file will be removed later anyways.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4004>