backlight/sysfs: Add sysfs based backlight support
The sysfs backlight is based on code from gnome-settings-daemon. It
pokes at the backlight interface of the sysfs. It uses udev to get
notified of updates and uses either a suid helper or a logind API to
write to the sysfs file.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4014>